enhanced stratified power

Equivalent geographical detector q-statistic under a spatial linear regression framework.

Usage

esp(
  formula,
  data,
  listw = NULL,
  yzone = NULL,
  discvar = "all",
  discnum = 3:8,
  model = "ols",
  durbin = FALSE,
  overlay = "and",
  alpha = 0.75,
  bw = "AIC",
  adaptive = TRUE,
  kernel = "gaussian",
  increase_rate = 0.05,
  cores = 1,
  ...
)

Arguments

formula

A formula for enhanced stratified power model.

data

An sf object of observation data. Please note that the column names of the independent variables should not be all or none.

listw

(optional) A list of listw. See spdep::mat2listw() and spdep::nb2listw() for details.

yzone

(optional) Spatial zones of the response variable. When yzone is NULL, the local zones q value is not calculated. Default is NULL.

discvar

(optional) Name of continuous variable columns that need to be discretized. Noted that when formula has discvar, data must have these columns. Default is all, which means all independent variables are used as discvar. When discvar is set to none, all independent variables do not need to be discretized.

discnum

(optional) Number of discretization. Default all will use 3:8.

model

(optional) The type of linear model used, default is ols. The model value must be any of ols, gwr, lag or error.

durbin

(optional) Whether to consider spatial durbin terms, default is false.

overlay

(optional) Spatial overlay method. One of and, or, intersection. Default is and.

alpha

(optional) Controlling the strength of spatial soft constraints, the larger the alpha, the stronger the spatial soft constraint. Default is 0.75.

bw

(optional) The bandwidth used in selecting models. The optimal bandwidth can be selected using one of two methods: AIC, and CV. Default will use AIC.

adaptive

(optional) Whether the bandwidth value is adaptive or not. Default is TRUE.

kernel

(optional) Kernel function. Default is gaussian.

increase_rate

(optional) The critical increase rate of the number of discretization. Default is 5%.

cores

(optional) Positive integer (default is 1). When cores are greater than 1, use multi-core parallel computing.

...

(optional) Other arguments passed to ClustGeo::hclustgeo().

Value

A list with espm class.

factor

global factor detection result

interaction

global interactive detection results

optdisc

independent variable optimal spatial discretization

localq

q values of explanatory variables under different response zones

zone

zones of the response variable

allfactor

factor detection results corresponding to different numbers of discreteization

model

regression model used to estimate equivalent q values

Examples

NTDs = sf::st_as_sf(gdverse::NTDs, coords = c('X','Y'))
g = esp(incidence ~ ., data = NTDs, discvar = 'none',
        model = 'ols', overlay = 'intersection', cores = 1)
g
#> ***           Enhanced Stratified Power     
#> 
#>  Q values are estimated using *Ordinary Least Square* 
#> 
#>  ------------- Global Power of Determinat : ------------
#> Variable  Qvalue AIC     BIC     LogLik  
#> watershed 0.638  -10.039 -10.039 15.020  
#> elevation 0.607  1.183   1.183   7.409   
#> soiltype  0.386  79.675  79.675  -33.838 
#> 
#>  ------------ Global Variable Interaction : ------------
#> Variable                Interaction  
#> watershed ∩ elevation Enhance, bi- 
#> watershed ∩ soiltype  Enhance, bi- 
#> elevation ∩ soiltype  Enhance, bi- 
#> 
#> ! Only the top ten items of global scale are displayed.
#> ! The others can be accessed through specific subsets.